split into components and create hook/context
